+++ index.d.ts
@@ -4,10 +4,23 @@ declare module 'human-ids' {
     adjective?: boolean;
     color?: boolean;
     noun?: boolean;
+    randomOrder?: boolean;
+    separator?: string;
     number?: {
       min?: number;
       max?: number;
       completeWithZeros?: boolean;
+    };
+    dictionary?: {
+      adjectives?: {
+        [key: string]: string[];
+      };
+      colors?: {
+        [key: string]: string[];
+      };
+      nouns?: {
+        [key: string]: string[];
+      }
     };
   }