+++ tests/index.test.js
@@ -2,12 +2,14 @@ const generateId = require('../index');
 
 describe('generateId', () => {
   test('should return a unique ID', () => {
-    const testLength = 1000;
+    const testLength = 1000000;
     const settings = {
       lang: 'en',
       adjective: true,
       color: true,
       noun: true,
+      randomOrder: true,
+      separator: '-',
       number: {
         min: 0,
         max: 9999,
@@ -20,6 +22,8 @@ describe('generateId', () => {
     // Generate testLength IDs
     for (let i = 0; i < testLength; i++) {
       const id = generateId(settings);
+
+      console.log(`${i} ID: ${id}`);
       ids.add(id);
     }